RFS sourceless: widen sidecar term-offset to int64
The terms.dat file can grow past 2 GiB on large segments (many long
unique terms), at which point the running currentTermOffset accumulator
in SidecarBuilder silently overflows int and is written as a negative
32-bit value into term-offsets.dat. SidecarReader would then treat that
sign-flipped value as a valid offset and read corrupt term bytes with
no exception surfaced.

Fix the on-disk format so term-offsets.dat stores little-endian int64
per term and track currentTermOffset as long in the builder. Reader
reads the offset via getLong instead of getInt. No production artifact
of the old 4-byte format exists yet — this PR introduced the package,
so there is nothing to migrate.

Test comments and the hand-authored large-file fixture are updated to
match the new 8-byte stride.

+++ SearchSnapshotExtractor/src/main/java/org/opensearch/migrations/bulkload/lucene/sidecar/SidecarBuilder.java
@@ -25,7 +25,7 @@ import lombok.extern.slf4j.Slf4j;
  * <p>Algorithm:
  * <ol>
  *   <li>{@link #registerTerm} streams term bytes to {@code terms.dat} and writes the
- *       LE int offset to {@code term-offsets.dat}.</li>
+ *       LE long offset to {@code term-offsets.dat}.</li>
  *   <li>{@link #accept} appends each {@code (docId, pos, termId)} triple as a fixed
  *       12-byte big-endian record into an in-memory sort buffer. Big-endian non-negative
  *       ints sort correctly under signed-int comparison, giving us {@code (docId ASC,
@@ -103,7 +103,7 @@ public final class SidecarBuilder implements PostingsSink, AutoCloseable {
     private final byte[] recordScratch = new byte[RECORD_BYTES];
 
     private int nextTermId = 0;
-    private int currentTermOffset = 0;
+    private long currentTermOffset = 0;
     private boolean built = false;
     private boolean closed = false;
 
@@ -137,10 +137,10 @@ public final class SidecarBuilder implements PostingsSink, AutoCloseable {
     @Override
     public int registerTerm(BytesRefLike term) throws IOException {
         int id = nextTermId++;
-        termOffsetsOut.writeInt(Integer.reverseBytes(currentTermOffset));
+        termOffsetsOut.writeLong(Long.reverseBytes(currentTermOffset));
         termsOut.writeInt(term.length());
         termsOut.write(term.bytes(), term.offset(), term.length());
-        currentTermOffset += 4 + term.length();
+        currentTermOffset += 4L + term.length();
         return id;
     }
